refactor: consolidate interfaces into domain/interfaces.py and remove dead code
Move IDocumentEngine, IAudioEngine, and ITextPipeline from their
implementation files into domain/interfaces.py where all other core
interfaces live. Remove six unused interfaces (IDocumentProcessor,
ITextProcessor, IAudioProcessor, IEngineCapabilityDetector,
IEnhancedTTSEngine, ITimestampedTTSEngine) and the SSMLCapability enum
that had no implementations anywhere in the codebase. Clean up
corresponding fake test helpers (FakeDocumentProcessor,
FakeAudioProcessor, FakeEngineCapabilityDetector) and update all imports
across 18 files.
